riscv: revert unnecessary type change

+++ riscv64/src/main.rs
@@ -24,8 +24,8 @@ use port::fdt::DeviceTree;
 core::arch::global_asm!(include_str!("l.S"));
 
 #[no_mangle]
-pub extern "C" fn main9(hartid: usize, dtb_ptr: usize) -> ! {
-    let dt = unsafe { DeviceTree::from_u64(dtb_ptr as u64).unwrap() };
+pub extern "C" fn main9(hartid: usize, dtb_ptr: u64) -> ! {
+    let dt = unsafe { DeviceTree::from_u64(dtb_ptr).unwrap() };
 
     // devcons::init(&dt);
     devcons::init_sbi();
